feat: move switch weekend mode to settings
