+++ src/components/Settings.svelte
@@ -1,9 +1,9 @@
 <script>
   import { slide } from 'svelte/transition'
-  import { boolSaveToLocal } from '../lib/store'
+  import { onMount } from 'svelte'
+  import { boolSaveToLocal, boolWeekendMode, weekdaysStore } from '../lib/store'
+  import { weekdays, weekend } from '../lib/constant'
 
-  export let boolWeekendMode
-  export let switchWeekendMode
 
   let boolMenu = false
   function toggleMenu() {
@@ -17,9 +17,27 @@
     }
   }
 
+  function switchWeekendMode() {
+    $boolWeekendMode = !$boolWeekendMode
+    localStorage.setItem('weekendMode', String($boolWeekendMode))
+    setWeekendMode()
+  }
+
+  function setWeekendMode() {
+    if ($boolWeekendMode) {
+      $weekdaysStore = [...$weekdaysStore, ...weekend]
+    } else {
+      $weekdaysStore = weekdays
+    }
+  }
+
   function switchLocalMode() {
     $boolSaveToLocal = !$boolSaveToLocal
   }
+
+  onMount(() => {
+    setWeekendMode()
+  })
 </script>
 
 <svelte:window on:keydown={handleKeydown} />
@@ -31,7 +49,7 @@
   <div transition:slide class="settings-menu">
     <ul>
       <li class="hover:bg-gray-200">
-        <div on:click={switchWeekendMode} class="checkbox {boolWeekendMode? 'checked' : ''}">
+        <div on:click={switchWeekendMode} class="checkbox {$boolWeekendMode? 'checked' : ''}">
           周末模式
         </div>
       </li>
